# src/modules/api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Salina ERP API
  version: "1.0.0"
  description: |
    REST API for Salina ERP - Publisher Royalty Management Platform.

    ## Authentication
    All API requests require authentication via Bearer tokens. See the Authentication section for details.

    ## Rate Limits
    - Default: 100 requests/minute + 1000 requests/hour
    - Premium: 1000 requests/minute + 10000 requests/hour
    - Auth endpoint: 10 requests/minute (IP-based)

    Rate limit headers are included in all responses.
  contact:
    name: Salina Support
    email: support@salina.app
  license:
    name: Proprietary
    url: https://salina.app/terms
servers:
  - url: https://{tenant}.salina.app/api/v1
    description: Production API
    variables:
      tenant:
        default: demo
        description: Your tenant subdomain
...
security:
  - bearerAuth: []
paths:
...
  /webhooks:
    get:
      tags: [Webhooks]
      summary: List webhook subscriptions
      responses:
        "200":
          description: List of webhook subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListResponse"
    post:
      tags: [Webhooks]
      summary: Create webhook subscription
      description: |
        Subscribe to receive webhook notifications for specified events.
        Requires `admin` scope. The signing secret is returned only once
        in the creation response - store it securely for signature verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookCreate"
      responses:
        "201":
          description: Webhook subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained via /auth/token endpoint.
        Include in Authorization header: `Bearer <token>`
...
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum: [title.created, title.updated, sale.created, statement.generated, onix.exported]
        is_active:
          type: boolean
        last_delivery_at:
          type: string
          format: date-time
        last_delivery_status:
          type: string
          enum: [success, failed]
        consecutive_failures:
          type: integer
        created_at:
          type: string
          format: date-time
    WebhookCreate:
      type: object
      required: [name, url, events]
      properties:
        name:
          type: string
          maxLength: 100
        url:
          type: string
          format: uri
          description: HTTPS URL to receive webhook payloads
        events:
          type: array
          minItems: 1
          items:
            type: string
            enum: [title.created, title.updated, sale.created, statement.generated, onix.exported]
...
    WebhookResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Webhook"
    WebhookListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"
...